Fix SocketException thrown by BSONObjectID.generate
When looking for network interfaces, openjdk parses /proc/net/if_inet6
on linux to find out which network interfaces are available.
Unfornutaly, the if_index might goes up to 2^32 not up to 255
values like the openjdk expects.

NetworkInterface.getNetworkInterfaces is a dangerous portion of code
and we should not call it if possible.

Further reference:
 * https://github.com/openjdk-mirror/jdk7u-jdk/blob/feeaec0647609a1e6266f902de426f1201f77c55/src/solaris/native/java/net/NetworkInterface.c#L1130
 * http://lxr.free-electrons.com/source/net/ipv6/addrconf.c?v=3.11#L3442
 * http://lxr.free-electrons.com/source/include/linux/netdevice.h?v=3.11#L1130
 * http://bugs.sun.com/bugdatabase/view_bug.do?bug_id=7078386
 * http://hg.openjdk.java.net/jdk8/tl/jdk/rev/b1814b3ea6d3