distinct: don't enforce <: Set[_] as the returned collection
Migrating to 0.12.0 and enjoying the higher-level `distinct` function.
However more often than not, all I want to do is iterate on the
values returned by `distinct`. In this case, a `Set` data
structure is suboptimal, because it does hashing.
I understand `distinct` returns unique values and therefore **can**
return a `Set`. But the collection type should depend on the way
the collection is used, and ReactiveMongo shouldn't make asumption
about that.