Use `System.nanoTime()` instead of `System.currentTimeMillis()`
Using the latter for measuring elapsed times is not a good idea since it relates
to a fixed point in time and takes adjustments of the system clock into
account (clock skew, NTP updates) so time may seem "to be running backwards".

The `System.nanoTime()` clock does not have this problem, at least since Java 8,
and is monotonically increasing (when supported by the OS).