Fix random connection issues
On fast machines with lots of cores, the Channel initialization function was often unreliable; channelInit() was being called at the same time that the channel properties were being added as attributes.

Fix by creating an actor to hold the properties for new channels, and stash requests for a channel's properties until after they've been added to the temporary cache.